requirements.txt
altair==5.4.1
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2021.10.8
cffi==1.15.0
channels==4.1.0
channels-redis==4.2.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
cssselect2==0.7.0
defusedxml==0.7.1
diff-match-patch==20230430
distro==1.9.0
Django==4.2.16
django-allauth==0.45.0
django-appconf==1.0.6
django-ckeditor==6.7.1
django-cpf-cnpj==1.0.0
django-crispy-forms==1.13.0
django-datatables-view==1.20.0
django-import-export==3.3.6
django-js-asset==2.2.0
django-json-widget==2.0.1
django-select2==8.1.2
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
dnspython==2.6.1
docopt==0.6.2
et-xmlfile==1.1.0
exceptiongroup==1.2.0
Faker==30.1.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.0
html5lib==1.1
idna==3.3
importlib_metadata==8.5.0
iniconfig==2.0.0
Jinja2==3.1.4
JPype1==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lxml==4.9.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupPy==1.14
MarkupSafe==2.1.5
mdurl==0.1.2
msgpack==1.1.0
narwhals==1.6.4
num2words==0.5.13
numpy==1.26.3
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.1.2
oscrypto==1.3.0
packaging==23.2
pandas==2.2.0
pdfminer.six==20221105
pdfplumber==0.10.3
pillow==10.2.0
pluggy==1.4.0
protobuf==5.28.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyarrow==17.0.0
pycparser==2.20
pydeck==0.9.1
Pygments==2.18.0
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
PyJWT==2.2.0
pymongo==4.8.0
PyNFe==0.4.4
pyOpenSSL==24.2.1
pypdf==4.0.0
PyPDF2==3.0.1
pypdfium2==4.26.0
pypng==0.20220715.0
pytest==8.1.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-stdnum==1.20
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0.1
qrcode==7.4.2
redis==5.1.0
referencing==0.35.1
reportlab==4.0.9
requests==2.31.0
requests-oauthlib==1.3.0
rich==13.8.1
rpds-py==0.20.0
signxml==3.2.1
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.34
sqlparse==0.4.2
streamlit==1.38.0
svglib==1.5.1
tablib==3.5.0
tenacity==8.5.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2023.4
tzlocal==5.2
uritools==4.0.2
urllib3==1.26.14
validate-docbr==1.10.0
watchdog==4.0.2
webencodings==0.5.1
xhtml2pdf==0.2.14
xlrd==2.0.1
xlwt==1.3.0
xmltodict==0.13.0
zipp==3.20.2
